+++ manager/manager/launcher/launcher_gazebo_view.py
@@ -1,12 +1,13 @@
-"""Launcher for Gazebo View"""""
-import time
 from src.manager.manager.launcher.launcher_interface import ILauncher
 from src.manager.manager.docker_thread.docker_thread import DockerThread
-
+from src.manager.manager.vnc.vnc_server import Vnc_server
+import time
+import os
+import stat
 
 
 class LauncherGazeboView(ILauncher):
-    """Launcher for Gazebo View"""
+    exercise_id: str
     display: str
     internal_port: str
     external_port: str
@@ -16,34 +17,42 @@ class LauncherGazeboView(ILauncher):
     threads = []
 
     def run(self, callback):
-        xserver_cmd = f"/usr/bin/Xorg -noreset +extension GLX +extension RANDR +extension RENDER -logfile ./xdummy.log -config ./xorg.conf {self.display}"
-        xserver_thread = DockerThread(xserver_cmd)
-        xserver_thread.start()
-        self.threads.append(xserver_thread)
-        time.sleep(0.1)
-        # Start VNC server without password, forever running in background
-        x11vnc_cmd = f"x11vnc -display {self.display} -nopw -forever -xkb -bg -rfbport {self.internal_port}"
-        x11vnc_thread = DockerThread(x11vnc_cmd)
-        x11vnc_thread.start()
-        self.threads.append(x11vnc_thread)
-
-        # Start noVNC with default port 6080 listening to VNC server on 5900
-        novnc_cmd = f"/noVNC/utils/launch.sh --listen {self.external_port} --vnc localhost:{self.internal_port}"
-        novnc_thread = DockerThread(novnc_cmd)
-        novnc_thread.start()
-        self.threads.append(novnc_thread)
+        DRI_PATH = os.path.join(
+            "/dev/dri", os.environ.get("DRI_NAME", "card0"))
+        ACCELERATION_ENABLED = self.check_device(DRI_PATH)
 
         # Configure browser screen width and height for gzclient
         gzclient_config_cmds = f"echo [geometry] > ~/.gazebo/gui.ini; echo x=0 >> ~/.gazebo/gui.ini; echo y=0 >> ~/.gazebo/gui.ini; echo width={self.width} >> ~/.gazebo/gui.ini; echo height={self.height} >> ~/.gazebo/gui.ini;"
-        time.sleep(0.1)
-        # Write display config and start gzclient
-        gzclient_cmd = (
-            f"export DISPLAY=:0;  {gzclient_config_cmds} gzclient --verbose")
+        gz_vnc = Vnc_server()
+
+        if ACCELERATION_ENABLED:
+            gz_vnc.start_vnc_gpu(self.display, self.internal_port, self.external_port, DRI_PATH)
+            # Write display config and start gzclient
+            gzclient_cmd = (
+                f"export DISPLAY=:0; {gzclient_config_cmds} export VGL_DISPLAY={DRI_PATH}; vglrun gzclient --verbose")
+        else:
+            gz_vnc.start_vnc(self.display, self.internal_port, self.external_port)
+            # Write display config and start gzclient
+            gzclient_cmd = (
+                f"export DISPLAY=:0; {gzclient_config_cmds} gzclient --verbose")
+
+        # wait for vnc and gazebo servers to load properly
+        if (self.exercise_id == "follow_person_newmanager"):
+            time.sleep(6)
+        else:
+            time.sleep(0.1)
+
         gzclient_thread = DockerThread(gzclient_cmd)
         gzclient_thread.start()
         self.threads.append(gzclient_thread)
 
         self.running = True
+
+    def check_device(self, device_path):
+        try:
+            return stat.S_ISCHR(os.lstat(device_path)[stat.ST_MODE])
+        except:
+            return False
 
     def is_running(self):
         return self.running
