+++ manager/manager/launcher/launcher_robot_display_view.py
@@ -1,6 +1,9 @@
 from src.manager.manager.launcher.launcher_interface import ILauncher
 from src.manager.manager.docker_thread.docker_thread import DockerThread
+from src.manager.manager.vnc.vnc_server import Vnc_server
 import time
+import os
+import stat
 
 
 class LauncherRobotDisplayView(ILauncher):
@@ -10,34 +13,43 @@ class LauncherRobotDisplayView(ILauncher):
     height: int
     width: int
     running = False
+    threads = []
 
     def run(self, callback):
-        xserver_cmd = f"/usr/bin/Xorg -noreset +extension GLX +extension RANDR +extension RENDER -logfile ./xdummy.log -config ./xorg.conf {self.display}"
-        xserver_thread = DockerThread(xserver_cmd)
-        xserver_thread.start()
-        time.sleep(0.1)
-        # Start VNC server without password, forever running in background
-        x11vnc_cmd = f"x11vnc -display {self.display} -nopw -forever -xkb -bg -rfbport {self.internal_port}"
-        x11vnc_thread = DockerThread(x11vnc_cmd)
-        x11vnc_thread.start()
+        DRI_PATH = os.path.join("/dev/dri", os.environ.get("DRI_NAME", "card0"))
+        ACCELERATION_ENABLED = self.check_device(DRI_PATH)
 
-        # Start noVNC with default port 2303 listening to VNC server on 5902
-        novnc_cmd = f"/noVNC/utils/launch.sh --listen {self.external_port} --vnc localhost:{self.internal_port}"
-        novnc_thread = DockerThread(novnc_cmd)
-        novnc_thread.start()
+        robot_display_vnc = Vnc_server()
+        
+        if (ACCELERATION_ENABLED):
+            robot_display_vnc.start_vnc_gpu(self.display, self.internal_port, self.external_port,DRI_PATH)
+            # Write display config and start the console
+            console_cmd = f"export VGL_DISPLAY={DRI_PATH}; export DISPLAY={self.display}; /usr/bin/Xorg -noreset +extension GLX +extension RANDR +extension RENDER -logfile ./xdummy.log -config ./xorg.conf {self.display}"
+        else:
+            robot_display_vnc.start_vnc(self.display, self.internal_port, self.external_port)
+            # Write display config and start the console
+            console_cmd = f"export DISPLAY={self.display};/usr/bin/Xorg -noreset +extension GLX +extension RANDR +extension RENDER -logfile ./xdummy.log -config ./xorg.conf {self.display}"
 
-        # Write display config and start RobotDisplayViewVNC
-        robot_cmd = (f"export DISPLAY=2;")
-        robot_thread = DockerThread(robot_cmd)
+        console_thread = DockerThread(console_cmd)
+        console_thread.start()
+        self.threads.append(console_thread)
 
-        robot_thread.start()
-        self.running = True
+        self.running = True        
+
+    def check_device(self, device_path):
+        try:
+            return stat.S_ISCHR(os.lstat(device_path)[stat.ST_MODE])
+        except:
+            return False
 
     def is_running(self):
         return self.running
 
     def terminate(self):
-        pass
+        for thread in self.threads:
+            thread.terminate()
+            thread.join()
+        self.running = False
 
     def died(self):
         pass
